+++ src/main/java/com/pika/Onlinegrocerystore/Services/OrderService.java
@@ -5,11 +5,11 @@ import com.pika.Onlinegrocerystore.domain.Order;
 import java.util.List;
 
 public interface OrderService {
-    Order findById(int id);
+    Order findById(Long id);
     List<Order> findAll();
     void update(Order customer);
     void create(Order customer);
-    void deleteById(int id);
-    void deleteByName(int id);
+    void deleteById(Long id);
+    void deleteByName(Long id);
     List<Order> findByName(String name);
 }
